requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.38.36
botocore==1.38.36
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
deprecation==2.1.0
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-extra-fields==3.7.0
django-filter==25.1
environ==1.0
filetype==1.2.0
frozenlist==1.7.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
jiter==0.10.0
jmespath==1.0.1
msgpack==1.1.1
multidict==6.4.4
ngrok==1.4.0
openai==1.90.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyOpenSSL==25.1.0
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-decouple==3.8
realtime==2.4.3
redis==6.2.0
requests==2.32.4
rest-framework-simplejwt==0.0.2
s3transfer==0.13.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
stripe==12.2.0
supabase==2.15.3
supafunc==0.9.4
tqdm==4.67.1
Twisted==25.5.0
txaio==23.1.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websockets==14.2
yarl==1.20.1
zope.interface==7.2
pandas==2.3.0
scikit_learn==1.7.0
gradio_client 
gunicorn==23.0.0